+++ src/lib/sortByText.js
@@ -1,0 +1,14 @@
+const sortByText = (arr, key) =>
+  arr.sort((a, b) => {
+    const nameA = a[key].toUpperCase();
+    const nameB = b[key].toUpperCase();
+    if (nameA < nameB) {
+      return -1;
+    }
+    if (nameA > nameB) {
+      return 1;
+    }
+    return 0;
+  });
+
+export default sortByText;