+++ src/modules/kiosks/components/DetailsHeader.js
@@ -1,18 +1,23 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Grid, Header, Icon } from 'semantic-ui-react';
+import CellHeartbeat from './CellHeartbeat';
 
-const DetailsHeader = ({ name, temp, connection, doorStatus }) => {
+const DetailsHeader = ({ name, temp, doorStatus, temperature }) => {
   const tempText = Number.isNaN(temp) ? '' : ` ${temp}°C`;
   return (
     <Grid>
       <Grid.Row relaxed="very" columns={4}>
-        <Grid.Column width={9}>
+        <Grid.Column width={6}>
           <Header as="h3">{name}</Header>
         </Grid.Column>
         <Grid.Column width={3}>
           Temp:
-          <b className="textGreen">{tempText}</b>
+          <b className="textRed">{tempText}</b>
+        </Grid.Column>
+        <Grid.Column width={3}>
+          Status:
+          <CellHeartbeat temperature={temperature} showTime={false} boldFont />
         </Grid.Column>
         <Grid.Column>
           Door:&nbsp;